supprimer un client et ajouter un produit a la base faits

+++ codes/src/Controller/Utilisateurs/AdminController.php
@@ -43,9 +43,17 @@ class AdminController extends AbstractController
      */
     public function deleteAccountAction(Utilisateurs $user): Response
     {
-        $this->em->remove($user);
-        $this->em->flush();
+        if ($user === $this->user)
+        {
+            $this->addFlash('info', "Sorry, you can't delete yourself");
+        }
+        else
+            {
+            $this->em->remove($user);
+            $this->em->flush();
 
+            $this->addFlash('info', "Customer deleted");
+            }
         return $this->redirectToRoute("main_index");
     }
 }
